feat(agent): gracefully shutdown snarkos node on exit
