+++ .vscode/settings.json
@@ -1,7 +1,4 @@
 {
-	"rust-analyzer.rustfmt.extraArgs": [
-		"+nightly"
-	],
-	"rust-analyzer.showUnlinkedFileNotification": false,
-	"rust-analyzer.check.extraArgs": ["--all-features", "--", "-D", "warnings", "--no-deps"]
+  "rust-analyzer.rustfmt.extraArgs": ["+nightly"],
+  "rust-analyzer.showUnlinkedFileNotification": false
 }
